requirements.txt
numpy
torch
torchvision
Pillow
scipy
click
tqdm
requests
flask
matplotlib
Image
